requirements.txt
astroid==2.4.2
beautifulsoup4==4.9.1
certifi==2020.6.20
cffi==1.14.0
chardet==3.0.4
colorama==0.4.3
cryptography==2.9.2
decorator==4.4.2
dnspython==1.16.0
emoji==0.5.4
flake8==3.8.3
idna==2.10
importlib-metadata==1.7.0
isort==4.3.21
lazy-object-proxy==1.4.3
mccabe==0.6.1
py-translate==1.0.3
pycodestyle==2.6.0
pycparser==2.20
pyflakes==2.2.0
pylint==2.5.3
pymongo==3.11.2
python-telegram-bot==12.8
requests==2.24.0
six==1.15.0
soupsieve==2.0.1
toml==0.10.1
tornado==6.0.4
typed-ast==1.4.1
urllib3==1.25.10
wrapt==1.12.1
zipp==3.1.0
